+++ pages/trades/[id]/index.tsx
@@ -4,10 +4,10 @@ import { getSession } from 'next-auth/client'
 import { GetServerSideProps } from 'next'
 
 // Layouts
-import { App as AppLayout } from './../../../layouts'
+import { App as AppLayout } from 'layouts'
 
 // Views
-import { TradeDetails as TradeDetailsView } from './../../../views'
+import { TradeDetails as TradeDetailsView } from 'views'
 
 function Dashboard({ toggleDarkMode, themeType }: Props): JSX.Element {
   return (
@@ -21,7 +21,7 @@ export const getServerSideProps: GetServerSideProps = async ctx => {
   const session = await getSession(ctx)
 
   if (!session && ctx.res) {
-    ctx.res.writeHead(302, { Location: '/api/auth/signin' })
+    ctx.res.writeHead(302, { Location: '/' })
     ctx.res.end()
   }
