+++ views/index.ts
@@ -1,2 +1,3 @@
+export { default as Home } from './Home'
 export { default as TradeDetails } from './TradeDetails'
 export { default as TradeFeed } from './TradeFeed'
